+++ src/field_data.h
@@ -37,16 +37,16 @@ class Field_Data
 		// ==================================
 		Field_Data(const Field_Data &rhs);
 		Field_Data(const SimParams &simparams);
-		Field_Data(long _x_pts, long _y_pts, long _z_pts, double _x_edge_mag, double _y_edge_mag, double _z_edge_mag);
+		Field_Data(int _x_pts, int _y_pts, int _z_pts, double _x_edge_mag, double _y_edge_mag, double _z_edge_mag);
 		~Field_Data();
 
 		// ==================================
 		// Data members
 		// ==================================
-		const long x_pts;
-		const long y_pts;
-		const long z_pts;
-		const long n_pts;
+		const int x_pts;
+		const int y_pts;
+		const int z_pts;
+		const long long n_pts;
 
 		const double x_edge_mag;
 		const double y_edge_mag;
@@ -67,25 +67,25 @@ class Field_Data
 		// ==================================
 		// Methods
 		// ==================================
-		long _index(long i, long j, long k) const;
+		long long _index(int i, int j, int k) const;
 
-		double &Ex_ind(long i, long j, long k);
-		double &Ex_ind(long ind);
+		double &Ex_ind(int i, int j, int k);
+		double &Ex_ind(long long ind);
 
-		double &Ey_ind(long i, long j, long k);
-		double &Ey_ind(long ind);
+		double &Ey_ind(int i, int j, int k);
+		double &Ey_ind(long long ind);
 
-		double &Ez_ind(long i, long j, long k);
-		double &Ez_ind(long ind);
+		double &Ez_ind(int i, int j, int k);
+		double &Ez_ind(long long ind);
 
-		double &Bx_ind(long i, long j, long k);
-		double &Bx_ind(long ind);
+		double &Bx_ind(int i, int j, int k);
+		double &Bx_ind(long long ind);
 
-		double &By_ind(long i, long j, long k);
-		double &By_ind(long ind);
+		double &By_ind(int i, int j, int k);
+		double &By_ind(long long ind);
 
-		double &Bz_ind(long i, long j, long k);
-		double &Bz_ind(long ind);
+		double &Bz_ind(int i, int j, int k);
+		double &Bz_ind(long long ind);
 
 		// ==================================
 		// Operators
